+++ src/components/CardTitle/CardTitle.tsx
@@ -1,12 +1,13 @@
 import { Slot, component$ } from '@builder.io/qwik';
+import { classNames } from '~/utils';
 
 type CardTitleProps = {
-  styling?: string;
+  class?: string;
 };
 
-export const CardTitle = component$<CardTitleProps>(({ styling }) => {
+export const CardTitle = component$<CardTitleProps>((props) => {
   return (
-    <p class={['italic text-sm font-heading mb-3', styling ?? ''].join(' ')}>
+    <p class={classNames('italic text-sm font-heading mb-3', props.class)}>
       <Slot />
     </p>
   );
