--- src/components/TalkCard/index.ts
@@ -1,1 +1,0 @@
-export { TalkCard } from './TalkCard';
