+++ src/components/TalkCardFix/index.ts
@@ -1,0 +1,1 @@
+export { TalkCard } from './TalkCardBroken';
