+++ src/components/builder-registry.ts
@@ -1,7 +1,7 @@
 import type { RegisteredComponent } from '@builder.io/sdk-qwik';
 
 import { TalkCardRegistry, TimelineRegistry } from './';
-import { RegisteredTalkCard } from './TalkCard/TalkCard.registered';
+import { RegisteredTalkCard } from './TalkCardFix/TalkCard.registered';
 
 export const CUSTOM_COMPONENTS: RegisteredComponent[] = [
   RegisteredTalkCard,
