+++ src/components/index.ts
@@ -1,11 +1,10 @@
-export { Card } from "./Card/Card";
-export { CardTitle } from "./CardTitle/CardTitle";
-export { CommonCard } from "./CommonCard/CommonCard";
-export { Footer } from "./Footer/Footer";
-export { Header } from "./Header/Header";
-export { Card } from "./layout/Card";
-export { RouterHead } from "./router-head/RouterHead";
-export { RegisteredTalkCard } from "./TalkCard/TalkCard.registered";
-export { TalkCard, TalkCardRegistry } from "./TalkCard/TalkCard";
-export { Timeline, TimelineRegistry } from "./timeline/Timeline";
-export { Heading } from "./typography/Heading";
+export { Card } from './Card/Card';
+export { CardTitle } from './CardTitle/CardTitle';
+export { CommonCard } from './CommonCard/CommonCard';
+export { Footer } from './Footer/Footer';
+export { Header } from './Header/Header';
+export { RouterHead } from './router-head/RouterHead';
+export { RegisteredTalkCard } from './TalkCardFix/TalkCard.registered';
+export { TalkCard, TalkCardRegistry } from './TalkCardFix/TalkCardBroken';
+export { Timeline, TimelineRegistry } from './Timeline/Timeline';
+export { Heading } from './typography/Heading';
